feat: #12 move BootstrapUseCaseImpl
